authentification web started

+++ Inocrea.CodaBox.ApiModel/ViewModel/StatementAccountViewModel.cs
@@ -3,15 +3,16 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Runtime.Serialization;
 using System.Text;
+using Inocrea.CodaBox.ApiModel.Models;
 
 namespace Inocrea.CodaBox.ApiModel.ViewModel
 {
     public class StatementAccountViewModel
     {
         [DataMember]
-        public decimal InitialBalance { get; set; }
+        public double InitialBalance { get; set; }
         [DataMember]
-        public decimal NewBalance { get; set; }
+        public double NewBalance { get; set; }
         [DataMember]
         public string InformationalMessage { get; set; }
         [DataMember]
